modified:   abstract_pynn_meanfield_model_standard.py 	modified:   builds/meanfield_base.py 	new file:   neuron_models/p_fit_polynomial.py 	modified:   neuron_models/params_from_network.py
add an other struct file to implemente params from fit of single neuron for
Vthre polynome
